refactor(frontend): generalize info banner
