+++ src/frontend/src/icp/components/info/Info.svelte
@@ -5,13 +5,12 @@
 		tokenCkEthLedger
 	} from '$icp/derived/ic-token.derived';
 	import InfoBitcoin from '$icp/components/info/InfoBitcoin.svelte';
-	import { type HideInfoKey, saveHideInfo, shouldHideInfo } from '$icp/utils/ck.utils';
-	import InfoBox from '$icp/components/info/InfoBox.svelte';
-	import { isNullish, nonNullish } from '@dfinity/utils';
+	import type { HideInfoKey } from '$icp/utils/ck.utils';
 	import InfoEthereum from '$icp/components/info/InfoEthereum.svelte';
 	import { isNetworkIdBTCMainnet, isNetworkIdETHMainnet } from '$icp/utils/ic-send.utils';
 	import type { OptionIcCkToken } from '$icp/types/ic';
 	import { token } from '$lib/stores/token.store';
+	import InfoBanner from '$lib/components/info/InfoBanner.svelte';
 
 	let mainnet = true;
 	$: mainnet =
@@ -35,27 +34,14 @@
 			: ckErc20
 				? 'oisy_ic_hide_erc20_info'
 				: undefined;
-
-	let hideInfo = true;
-	$: hideInfo = nonNullish(key) ? shouldHideInfo(key) : true;
-
-	const close = () => {
-		hideInfo = true;
-
-		if (isNullish(key)) {
-			return;
-		}
-
-		saveHideInfo(key);
-	};
 </script>
 
 {#if ckBTC || ckETH || ckErc20}
-	<InfoBox {hideInfo} on:click={close}>
+	<InfoBanner {key}>
 		{#if ckBTC}
 			<InfoBitcoin />
 		{:else}
 			<InfoEthereum />
 		{/if}
-	</InfoBox>
+	</InfoBanner>
 {/if}
