+++ src/frontend/src/lib/components/info/InfoBox.svelte
